Abstract out FaradayVAE
This is prevent circular imports.
`src/opensynth/models/faraday/model.py` depends on `_lightning.py` scripts which themselves take FaradayVAE as input.
Before FaradayVAE was defined in `src/opensynth/models/faraday/model.py` creating a circular import. I have moved FaradayVAE to vae_models.py to prevent circular import.
