+++ src/opensynth/models/faraday/__init__.py
@@ -1,11 +1,7 @@
 # Copyright Contributors to the Opensynth-energy Project.
 # SPDX-License-Identifier: Apache-2.0
 
-from opensynth.models.faraday.model import (
-    Decoder,
-    Encoder,
-    FaradayModel,
-    FaradayVAE,
-)
+from opensynth.models.faraday.model import Decoder, Encoder, FaradayModel
+from opensynth.models.faraday.vae_model import FaradayVAE
 
 __all__ = ["FaradayVAE", "FaradayModel", "Encoder", "Decoder"]
